5692 - Bulk download: Clean up data fetching
