+++ src/server/controller/cycleData/getBulkDownload/_getData/_getTableData.ts
@@ -1,0 +1,25 @@
+import { CountryIso } from 'meta/area/countryIso'
+import { Assessment } from 'meta/assessment/assessment'
+import { Cycle } from 'meta/assessment/cycle'
+import { RecordAssessmentData } from 'meta/data/recordData'
+
+import { getLastPublishedData } from 'server/controller/cycleData/getLastPublishedData'
+import { getTableData } from 'server/controller/cycleData/getTableData'
+
+type Props = {
+  assessment: Assessment
+  countryISOs: Array<CountryIso>
+  cycle: Cycle
+  isLastPublishedCycle: boolean
+  tableNames: Array<string>
+}
+
+export const _getTableData = async (props: Props): Promise<RecordAssessmentData> => {
+  const { assessment, countryISOs, cycle, isLastPublishedCycle, tableNames } = props
+
+  if (isLastPublishedCycle) {
+    return getLastPublishedData({ assessment, countryISOs, tableNames })
+  }
+
+  return getTableData({ assessment, countryISOs, cycle, mergeOdp: true, tableNames })
+}
